L017 - current user & added security for all requests

+++ src/main/java/com/apeter/blog/article/service/ArticleApiService.java
@@ -1,6 +1,9 @@
 package com.apeter.blog.article.service;
 
 import com.apeter.blog.article.mapping.ArticleMapping;
+import com.apeter.blog.auth.exceptions.AuthException;
+import com.apeter.blog.auth.exceptions.NoAccessException;
+import com.apeter.blog.auth.service.AuthService;
 import com.apeter.blog.base.api.request.SearchRequest;
 import com.apeter.blog.base.api.response.SearchResponse;
 import com.apeter.blog.article.api.request.ArticleRequest;
@@ -8,11 +11,13 @@ import com.apeter.blog.article.exception.ArticleExistException;
 import com.apeter.blog.article.exception.ArticleNoExistException;
 import com.apeter.blog.article.model.ArticleDoc;
 import com.apeter.blog.article.repository.ArticleRepository;
+import com.apeter.blog.base.service.CheckAccess;
 import com.apeter.blog.user.exception.UserNoExistException;
 import com.apeter.blog.user.model.UserDoc;
 import com.apeter.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
@@ -25,17 +30,17 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class ArticleApiService {
+public class ArticleApiService extends CheckAccess<ArticleDoc> {
     private final ArticleRepository articleRepository;
     private final MongoTemplate mongoTemplate;
     private final UserRepository userRepository;
+    private final AuthService authService;
 
-    public ArticleDoc create(ArticleRequest request) throws ArticleExistException, UserNoExistException {
+    public ArticleDoc create(ArticleRequest request) throws AuthException {
 
-        Optional<UserDoc> userDoc = userRepository.findById(request.getOwnerId());
-        if(!userDoc.isPresent()) throw new UserNoExistException();
+        UserDoc userDoc = authService.currentUser();
 
-        ArticleDoc articleDoc = ArticleMapping.getInstance().getRequestMapping().convert(request);
+        ArticleDoc articleDoc = ArticleMapping.getInstance().getRequestMapping().convert(request, userDoc.getId());
         articleRepository.save(articleDoc);
         return articleDoc;
     }
@@ -66,15 +71,16 @@ public class ArticleApiService {
         return SearchResponse.of(articleDocs, count);
     }
 
-    public ArticleDoc update(ArticleRequest request) throws ArticleNoExistException {
+    public ArticleDoc update(ArticleRequest request) throws ArticleNoExistException, AuthException, NoAccessException {
         Optional<ArticleDoc> articleDocOptional = articleRepository.findById(request.getId());
         if (!articleDocOptional.isPresent()) {
             throw new ArticleNoExistException();
         }
 
         ArticleDoc oldDoc = articleDocOptional.get();
+        UserDoc owner = checkAccess(oldDoc);
 
-        ArticleDoc articleDoc = ArticleMapping.getInstance().getRequestMapping().convert(request);
+        ArticleDoc articleDoc = ArticleMapping.getInstance().getRequestMapping().convert(request, owner.getId());
         articleDoc.setId(request.getId());
         articleDoc.setOwnerId(oldDoc.getOwnerId());
         articleRepository.save(articleDoc);
@@ -82,7 +88,18 @@ public class ArticleApiService {
         return articleDoc;
     }
 
-    public void deleteById(ObjectId id) {
+    public void deleteById(ObjectId id) throws NoAccessException, AuthException, ChangeSetPersister.NotFoundException {
+        checkAccess(articleRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         articleRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(ArticleDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
