+++ src/main/java/com/apeter/blog/todoTask/service/TodoTaskApiService.java
@@ -1,7 +1,11 @@
 package com.apeter.blog.todoTask.service;
 
+import com.apeter.blog.auth.exceptions.AuthException;
+import com.apeter.blog.auth.exceptions.NoAccessException;
+import com.apeter.blog.auth.service.AuthService;
 import com.apeter.blog.base.api.request.SearchRequest;
 import com.apeter.blog.base.api.response.SearchResponse;
+import com.apeter.blog.base.service.CheckAccess;
 import com.apeter.blog.todoTask.api.request.TodoTaskRequest;
 import com.apeter.blog.todoTask.api.request.TodoTaskSearchRequest;
 import com.apeter.blog.todoTask.exception.TodoTaskExistException;
@@ -10,9 +14,11 @@ import com.apeter.blog.todoTask.mapping.TodoTaskMapping;
 import com.apeter.blog.todoTask.model.TodoTaskDoc;
 import com.apeter.blog.todoTask.repository.TodoTaskRepository;
 import com.apeter.blog.user.exception.UserNoExistException;
+import com.apeter.blog.user.model.UserDoc;
 import com.apeter.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
@@ -28,17 +34,17 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class TodoTaskApiService {
+public class TodoTaskApiService extends CheckAccess<TodoTaskDoc> {
     private final TodoTaskRepository todoTaskRepository;
     private final MongoTemplate mongoTemplate;
     private final UserRepository userRepository;
+    private final AuthService authService;
 
-    public TodoTaskDoc create(TodoTaskRequest request) throws TodoTaskExistException, UserNoExistException {
+    public TodoTaskDoc create(TodoTaskRequest request) throws AuthException {
 
-        if (userRepository.findById(request.getOwnerId()).isEmpty())
-            throw new UserNoExistException();
+        UserDoc userDoc = authService.currentUser();
 
-        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request);
+        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request, userDoc.getId());
         todoTaskRepository.save(todoTaskDoc);
         return todoTaskDoc;
     }
@@ -49,11 +55,10 @@ public class TodoTaskApiService {
 
     public SearchResponse<TodoTaskDoc> search(
             TodoTaskSearchRequest request
-    ) {
-        if(request.getOwnerId() == null)
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+    ) throws AuthException {
+        UserDoc userDoc = authService.currentUser();
 
-        Criteria criteria = Criteria.where("ownerId").is(request.getOwnerId());
+        Criteria criteria = Criteria.where("ownerId").is(userDoc.getId());
         if (request.getQuery() != null && !request.getQuery().equals("")) {
             criteria = criteria.orOperator(
                     Criteria.where("title").regex(request.getQuery(), "i")
@@ -71,15 +76,16 @@ public class TodoTaskApiService {
         return SearchResponse.of(todoTaskDocs, count);
     }
 
-    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNoExistException {
+    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNoExistException, NoAccessException, AuthException {
         Optional<TodoTaskDoc> todoTaskDocOptional = todoTaskRepository.findById(request.getId());
         if (!todoTaskDocOptional.isPresent()) {
             throw new TodoTaskNoExistException();
         }
 
         TodoTaskDoc oldTask = todoTaskDocOptional.get();
+        UserDoc userDoc = checkAccess(oldTask);
 
-        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request);
+        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request, userDoc.getId());
 
         todoTaskDoc.setId(request.getId());
         todoTaskDoc.setOwnerId(oldTask.getOwnerId());
@@ -88,7 +94,18 @@ public class TodoTaskApiService {
         return todoTaskDoc;
     }
 
-    public void deleteById(ObjectId id) {
+    public void deleteById(ObjectId id) throws NoAccessException, AuthException, ChangeSetPersister.NotFoundException {
+        checkAccess(todoTaskRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         todoTaskRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(TodoTaskDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
